updated CRUD functionality for AI and manual recipes
