+++ README.md
@@ -48,21 +48,27 @@ The schema also includes methods for password hashing, JWT token creation, and p
 It stores recipe information, including recipe name, ingredients, serving size, cooking instructions, nutrition info, meal preparation time, recipe category, special diet type, recipe tags, complexity level and meal image. 
 It uses subdocuments for recipe ingredients and tags.
 
-# MANUAL RECIPE ENTRY FUNCTIONALITY
+# Recipe CRUD (create, read, update, delete) Functionality
 
-The manual recipe entry functionality allows users to create, retrieve/read, update, and delete recipes manually within the app.
+An overview of the CRUD (Create, Read, Update, Delete) backend functionality for recipes.
+This functionality allows users to manage recipes, whether they are manually entered or generated by an AI assistant.
 
 Endpoints: 
 
 1. Endpoints:
-Create Manual Recipe: POST /api/v1/manual-recipe
-Get All Recipes: GET /api/v1/manual-recipe
-Get Manual Recipe by ID: GET /api/v1/manual-recipe/:recipeId
-Update Manual Recipe: PATCH /api/v1/manual-recipe/:recipeId
-Delete Manual Recipe: DELETE /api/v1/manual-recipe/:recipeId
+- Fetch AI Recipe. This endpoint fetches a recipe using the OpenAI API based on user input:                                                     POST /api/v1/recipes
+- Create AI Recipe. The created AI recipe is stored in the DB:                                                                                      POST /api/v1/recipes/add-ai
+- Create Manual Recipe. This endpoint allows authenticated users to manually create a new recipe. The created recipe is stored in the database: POST /api/v1/recipes/add-manual
+ -Get All Recipes. This endpoint retrieves all recipes created by the authenticated user, including both manually entered and AI-generated recipes: GET /api/v1/recipes
+- Get Recipe by ID. Retrieves a specific recipe by its ID, whether it was manually entered or AI-generated:                                         GET /api/v1/recipes/:recipeId
+- Update Recipe. Allows the authenticated user to update a specific recipe, whether it was manually entered or AI-generated.:                   PATCH /api/v1/recipes/:recipeId
+- Delete Recipe. Allows the authenticated user to delete a specific recipe, whether it was manually entered or AI-generated:                        DELETE /api/v1/recipes/:recipeId
 
 2. Authentication:
 Users must be authenticated to use these endpoints. Authentication tokens can be obtained by logging in. Users can only create, retrieve, update, and delete recipes that they have created.
 
-3. Error Handling
-The API follows RESTful conventions for error handling. When errors occur, appropriate HTTP status codes are returned along with error messages in the response body.
+4. User-Recipe association
+When a user creates a manual recipe or an AI-generated recipe, the 'recipeCreatedBy' field is set to the user's unique identifier. This ensures that each recipe is linked to the user who created it. When retrieving, updating, or deleting a recipe, the user's authentication token is used to identify the user. The user's token is verified, and the associated recipes are accessed or modified based on the user's authentication.
+
+5. Error Handling
+All responses follow a consistent format. When errors occur, appropriate HTTP status codes are returned along with error messages in the response body.
